pyproject.toml
[project]
name = "ivexes"
version = "0.1.0"
description = "A Python package for automating vulnerability analysis and exploit generation"
readme = "README.md"
license = "GPL-3.0-or-later"
dependencies = [
  "dotenv",
  "pydantic",
  "pydantic_settings",
  "paramiko",
  "chromadb",
  "pynvim",
  "docker",
  "chardet",
  "openai",
  "openai-agents",
  "python-dotenv",
  "click",
  "tiktoken",
  "mitreattack-python",
]

[project.optional-dependencies]
dev = [
  "pre-commit",
  "ruff",
]

[project.urls]
Repository = "https://github.com/LetsDrinkSomeTea/ivexes"

[tool.ruff]
extend-exclude = [".venv", "examples"]

[tool.ruff.format]
skip-magic-trailing-comma = true
quote-style = "single"
docstring-code-format = true

[tool.ruff.lint]
select = ["D"]

[tool.ruff.lint.pydocstyle]
convention = "google"
